% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmReference.R
\name{Binomial-ergmReference}
\alias{Binomial-ergmReference}
\alias{InitErgmReference.Binomial}
\title{Binomial-reference ERGM}
\usage{
# Binomial(trials)
}
\arguments{
\item{trails}{model parameter}
}
\description{
Specifies each dyad's baseline distribution to be binomial with
\code{trials} trials and success probability of \eqn{0.5} :
\eqn{h(y)=\prod_{i,j}{{\mathrm{trials}}\choose{y_{i,j}}}} . Using
\link[=ergmTerm]{valued ERGM terms} that are
"generalized" from their binary counterparts, with form
\code{"sum"} (see previous link for the list) produces logistic
regression.
}
\seealso{
\code{\link{ergmReference}} for index of reference distributions currently visible to the package.
}
\concept{bipartite}
\concept{directed}
\concept{discrete}
\concept{finite}
\concept{nonnegative}
\concept{undirected}
\concept{valued}
