% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmProposal.R
\name{Disc-ergmProposal}
\alias{Disc-ergmProposal}
\alias{InitWtErgmProposal.Disc}
\title{Sampling for some discrete-reference ERGMs}
\description{
This proposal implements \link{Poisson-ergmReference},
\link{Geometric-ergmReference}, \link{Binomial-ergmReference}, and
\link{DiscUnif-ergmReference} with arbitrary dyad level constraints.
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="Disc"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="Disc"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="Disc"))}}
}
\seealso{
\link{DiscTNT-ergmProposal}
}
\concept{bipartite}
\concept{directed}
\concept{discrete}
\concept{undirected}
\concept{valued}
