% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.count-package.R
\docType{package}
\name{ergm.count-package}
\alias{ergm.count-package}
\title{Fit, Simulate and Diagnose Exponential-Family Models for Networks with Count
Edges}
\description{
\code{\link[=ergm.count-package]{ergm.count}} is a set of extensions to
package \code{\link[=ergm-package]{ergm}} to fit and simulate from
exponential-family random graph models for networks whose edge weights are
counts. For a list of functions type \code{help(package='ergm')} and
\code{help(package='ergm.count')}
}
\details{

}
\references{
Handcock MS, Hunter DR, Butts CT, Goodreau SG, Krivitsky PN and
Morris M (2012). \emph{Fit, Simulate and Diagnose Exponential-Family Models for
Networks}. Version 3.1.  Project home page at <URL: https://www.statnet.org>,
<URL: CRAN.R-project.org/package=ergm>.

Krivitsky PN (2012). Exponential-Family Random Graph Models for Valued
Networks. \emph{Electronic Journal of Statistics}, 2012, 6, 1100-1128.
\doi{10.1214/12-EJS696}
}
\seealso{
\code{\link{ergm-terms}}, \code{\link{ergm-references}}
}
\author{
Pavel N. Krivitsky \email{pavel@statnet.org}
}
\keyword{models}
\keyword{package}
