% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.count-terms-index.R
\docType{package}
\name{ergmReference}
\alias{ergmReference}
\alias{ergm-references}
\alias{references-ergm}
\alias{ergm.references}
\alias{references.ergm}
\alias{Poisson}
\alias{Binomial}
\alias{Geometric}
\alias{InitWtErgmProposal.Poisson}
\alias{InitWtErgmProposal.ZIPoisson}
\alias{InitWtErgmProposal.PoissonTNT}
\alias{InitWtErgmProposal.PoissonNonObserved}
\alias{InitWtErgmProposal.Binomial}
\alias{InitWtErgmProposal.BinomialNonObserved}
\alias{InitWtErgmProposal.Geometric}
\alias{InitWtErgmProposal.GeometricNonObserved}
\title{Reference Measures for Exponential-Family Random Graph Models for Counts}
\description{
This page describes the possible reference measures (baseline distributions)
for modeling count data found in the \code{\link[=ergm.count-package]{ergm.count}} package.

Each of these is specified on the RHS of a one-sided formula passed as the
\code{reference} argument to \code{\link{ergm}}.  See the \code{\link{ergm}} documentation for a complete description of how
reference measures are specified.
}
\section{Known issues}{

\subsection{Parameter space constraints}{

Poisson- and geometric-reference ERGMs have an unbouded sample space. This
means that the parameter space may be constrained in complex ways that
depend on the terms used in the model. At this time \code{\link{ergm}} has
no way to detect when a parameter configuration had strayed outside of the
parameter space, but it may be noticeable on a runtime trace plot (activated
via \code{MCMC.runtime.traceplot} control parameter), when the simulated
values keep climbing upwards. (See Krivitsky (2012) for a further
discussion.)

A possible remedy if this appears to occur is to try lowering the control
parameter \code{MCMLE.steplength}.
}
}

\section{Possible reference measures to represent baseline distributions}{

\subsection{Reference index}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexLatex(ergm:::.buildTermsDataframe("ergmReference"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexText(ergm:::.buildTermsDataframe("ergmReference"))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexHtml(ergm:::.buildTermsDataframe("ergmReference"))}}
}
}

\references{
\itemize{
\item Krivitsky PN (2012). Exponential-Family Random Graph Models for
Valued Networks. \emph{Electronic Journal of Statistics}, 2012, 6,
1100-1128. \doi{10.1214/12-EJS696}
\item Shmueli G, Minka TP, Kadane JB, Borle S, and Boatwright P (2005). A Useful
Distribution for Fitting Discrete Data: Revival of the
Conway--Maxwell--Poisson Distribution. \emph{Journal of the Royal
Statistical Society: Series C}, 54(1): 127-142.
}
}
\seealso{
\code{\link[=ergm-package]{ergm}}, \code{\link{network}}, \verb{\\\%v\\\%}, \verb{\\\%n\\\%}, \code{sna}, \code{\link{summary.ergm}}, \code{\link{print.ergm}}
}
\keyword{models}
