% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm-terms-index.R
\docType{package}
\name{ergmTerm}
\alias{ergmTerm}
\alias{ergm-terms}
\alias{ergm.terms}
\alias{terms-ergm}
\alias{terms.ergm}
\title{Terms used in Exponential Family Random Graph Models Specific to Counts}
\description{
This page describes the possible terms (and hence network statistics)
included in the \code{\link[=ergm.count-package]{ergm.count}} package.

See the \code{\link[ergm:ergm-terms]{ergm-terms}} documentation in the \code{ergm}
package for a complete description of what ERGM terms are and how they are
used.
}
\section{Terms to represent network statistics included in the \code{\link[=ergm.count-package]{ergm.count}} package}{

\subsection{Term index}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexLatex(ergm:::.buildTermsDataframe("ergmTerm"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexText(ergm:::.buildTermsDataframe("ergmTerm"))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexHtml(ergm:::.buildTermsDataframe("ergmTerm"))}}
}

\subsection{Frequently-used terms}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixLatex(ergm:::.termMatrix("ergmTerm", categories=ergm:::FREQUENTLY_USED_TERM_CATEGORIES, only.include='frequently-used'))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixText(ergm:::.termMatrix("ergmTerm", categories=ergm:::FREQUENTLY_USED_TERM_CATEGORIES, only.include='frequently-used'))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixHtml(ergm:::.termMatrix("ergmTerm", categories=ergm:::FREQUENTLY_USED_TERM_CATEGORIES, only.include='frequently-used'))}}
}

\subsection{Operator terms}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixLatex(ergm:::.termMatrix("ergmTerm", categories=ergm:::OPERATOR_CATEGORIES, only.include='operator'))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixText(ergm:::.termMatrix("ergmTerm", categories=ergm:::OPERATOR_CATEGORIES, only.include='operator'))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixHtml(ergm:::.termMatrix("ergmTerm", categories=ergm:::OPERATOR_CATEGORIES, only.include='operator'))}}
}

\subsection{All terms}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixLatex(ergm:::.termMatrix("ergmTerm"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixText(ergm:::.termMatrix("ergmTerm"))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixHtml(ergm:::.termMatrix("ergmTerm"))}}
}

\subsection{Terms by concept}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatTocLatex(ergm:::.termToc("ergmTerm"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatTocText(ergm:::.termToc("ergmTerm"))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatTocHtml(ergm:::.termToc("ergmTerm"))}}
}
}

\references{
\itemize{
\item Handcock M. S., Hunter D. R., Butts C. T.,
Goodreau S. G., Krivitsky P. N. and Morris M. (2012). \emph{Fit, Simulate and
Diagnose Exponential-Family Models for Networks}. Version 3.1.  Project home
page at <URL: https://www.statnet.org>, <URL:
CRAN.R-project.org/package=ergm>.
\item Krivitsky P. N. (2012). Exponential-Family Random Graph Models for
Valued Networks. \emph{Electronic Journal of Statistics}, 2012, 6,
1100-1128. \doi{10.1214/12-EJS696}
\item Shmueli G., Minka T. P., Kadane J. B., Borle S., and Boatwright P.
(2005). A Useful Distribution for Fitting Discrete Data: Revival of the
Conway--Maxwell--Poisson Distribution. \emph{Journal of the Royal
Statistical Society: Series C}, 54(1): 127-142.
\item Shmueli G., Minka T. P., Kadane J. B., Borle S., and Boatwright P.
(2005). A Useful Distribution for Fitting Discrete Data: Revival of the
Conway--Maxwell--Poisson Distribution. \emph{Journal of the Royal
Statistical Society: Series C}, 54(1): 127-142.
\item Kadane, Joseph B.  (2016) Sums of Possibly Associated Bernoulli
Variables: The Conway-Maxwell-Binomial Distribution. \emph{Bayesian
Analysis}, 11(2): 403--420. \doi{10.1214/15-BA955}
}
}
\seealso{
\code{\link[ergm:ergm-terms]{ergm-terms}} (from the
\code{\link[=ergm-package]{ergm}} package), \code{\link{ergm}}, \code{\link{network}}, \verb{\%v\%}, \verb{\%n\%}
}
\keyword{models}
