% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmProposal.R
\name{Geometric-ergmProposal}
\alias{Geometric-ergmProposal}
\alias{InitWtErgmProposal.Geometric}
\title{TODO}
\description{
TODO
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="Geometric"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="Geometric"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="Geometric"))}}
}
\concept{bipartite}
\concept{directed}
\concept{discrete}
\concept{nonnegative}
\concept{undirected}
\concept{valued}
