% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.count-package.R
\docType{package}
\name{ergm.count-package}
\alias{ergm.count}
\alias{ergm.count-package}
\title{Fit, Simulate and Diagnose Exponential-Family Models for Networks with Count
Edges}
\description{
\code{\link[=ergm.count-package]{ergm.count}} is a set of extensions to
package \code{\link[=ergm-package]{ergm}} to fit and simulate from
exponential-family random graph models for networks whose edge weights are
counts \insertCite{Kr12e}{ergm.count}.
}
\details{
Mainly, it implements Poisson, binomial, geometric, and discrete
uniform dyadwise reference measures for valued ERGMs (documented
here in \code{\link{ergmReference}}), and provides some count-specific change
statistics (documented in \code{\link{ergmTerm}})
\insertCite{Kr12e,KrHu23e}{ergm.count}, including
\code{\link[=CMP-ergmTerm]{CMP}} for the Conway--Maxwell--Poisson Distribution
\insertCite{ShMi05u}{ergm.count}.

For a complete list of the functions, use \code{library(help="ergm")} and
\code{library(help="ergm.count")} or read the rest of the manual.

When publishing results obtained using this package, please cite the
original authors as described in \code{citation(package="ergm.count")}.

All programs derived from this package must cite it.

This package contains functions specific to using \code{\link{ergm}} to
model networks whose dyad values are counts. Examples include counts of
conversations, messages, and other interactions.

For detailed information on how to download and install the software, go to
the Statnet project website: \url{https://statnet.org}. A tutorial, support
newsgroup, references and links to further resources are provided there.
}
\section{Known issues}{

\subsection{Parameter space constraints}{

Poisson- and geometric-reference ERGMs have an unbouded sample space. This
means that the parameter space may be constrained in complex ways that
depend on the terms used in the model. At this time \code{\link{ergm}} has
no way to detect when a parameter configuration had strayed outside of the
parameter space, but it may be noticeable on a runtime trace plot (activated
via \code{MCMC.runtime.traceplot} control parameter), when the simulated
values keep climbing upwards. (See \insertCite{Kr12e;textual}{ergm.count} for a further
discussion.)

A possible remedy if this appears to occur is to try lowering the
\code{\link[=control.ergm]{control.ergm()}} parameter \code{MCMLE.steplength}.
}
}

\references{
\insertAllCited{}
}
\seealso{
\code{\link{ergmTerm}}, \code{\link{ergmReference}}
}
\author{
\strong{Maintainer}: Pavel N. Krivitsky \email{pavel@statnet.org} (\href{https://orcid.org/0000-0002-9101-3362}{ORCID})

Other contributors:
\itemize{
  \item Mark S. Handcock \email{handcock@stat.ucla.edu} [contributor]
  \item David R. Hunter \email{dhunter@stat.psu.edu} [contributor]
  \item Joyce Cheng \email{joyce.cheng@student.unsw.edu.au} [contributor]
}

}
