% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmReference.R
\name{Geometric-ergmReference}
\alias{Geometric-ergmReference}
\alias{InitErgmReference.Geometric}
\title{Geometric-reference ERGM}
\usage{
# Geometric
}
\description{
Specifies
each dyad's baseline distribution to be uniform on the natural
numbers (and \eqn{0} ): \eqn{h(y)=1} . In itself, this
"distribution" is improper, but in the presence of
\code{\link[=sum-ergmTerm]{sum}} , a geometric
distribution is induced. Using \code{\link[=CMP-ergmTerm]{CMP}} (in addition to
\code{\link[=sum-ergmTerm]{sum}} ) induces a
Conway-Maxwell-Poisson distribution that is geometric when its
coefficient is \eqn{0} and Poisson when its coefficient is
\eqn{-1} .
}
\seealso{
\code{\link{ergmReference}} for index of reference distributions currently visible to the package.
}
\concept{bipartite}
\concept{directed}
\concept{discrete}
\concept{nonnegative}
\concept{undirected}
\concept{valued}
